Desktop/Bonn/Autoscraper/requirements.txt
Django
beautifulsoup4
pandas
nltk
textblob
matplotlib
azure-storage-blob
